style/FE: Responsive header
